refactor bridge contract
